requirements.txt
Flask==1.0.2
Flask-JWT-Extended==3.14.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
slackclient==1.3.0